Moves Piece to 'model' package
